benchlab/fastapi/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.24.0
pyserial>=3.5
python-dotenv==1.1.1
